requirements.txt
tensorflow-gpu==1.12.2
numpy==1.14.5
nibabel==2.5.1
scipy==1.3.1
scikit-image==0.15.0
scikit-learn==0.21.3
opencv-python==4.1.1.26
tensorboard==1.12.2
nipy==0.4.2
